Simplify sanitizing identifiers
